% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bintanja_vandewal_2008}
\alias{bintanja_vandewal_2008}
\title{Default Sea Level Reconstruction Dataset (Bintanja & van de Wal, 2008)}
\format{
\subsection{\code{bintanja_vandewal_2008}}{

A data frame with 30,002 rows and 3 columns:
\describe{
\item{Time}{Time in the past (kya)}
\item{Sealevel_Uncorrected}{Raw sea level estimate from original dataset
(metres)}
\item{Sealevel_Corrected}{Sea level estimate, with present day set to 0
(metres)}
}
}
}
\source{
\url{https://doi.org/10.1038/nature07158}
}
\usage{
bintanja_vandewal_2008
}
\description{
Sea level reconstruction file from Bintanja & van de Wal (2008)
}
\keyword{datasets}
