% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makerasters.R
\name{makemaps}
\alias{makemaps}
\title{Generate maps of Pleistocene island extents}
\usage{
makemaps(inputraster, epsg, intervalfile = "output/intervals.csv")
}
\arguments{
\item{inputraster}{An input bathymetry raster in ASCII (.asc) format. Although PleistoDist should theoretically be able to
use any time of ASCII-formatted bathymetry grid as input, this tool has been tested specifically with data from the General
Bathymetric Chart of the Oceans (GEBCO: https://www.gebco.net). Locality-specific bathymetric maps can be downloaded from
https://download.gebco.net/.}

\item{epsg}{The projected coordinate system in EPSG code format. Because of the curvature of the Earth's surface, we need to apply a map
projection to accurately calculate straight-line distances between points instead of using the default WGS84 geographical coordinate system.
Users should specify a projected coordinate system appropriate to the geographic region being analysed using the projection's
associated EPSG code (https://epsg.org/home). Geographic coordinate system projections are not recommended as those will result
in distance matrices calculated in decimal degrees rather than in distance units.}

\item{intervalfile}{This is the master control file generated using either the getintervals_time() or getintervals_sealvl()
function. By default, this function will use the "intervals.csv" file stored in the output folder, but users can also specify their
own custom interval file (with nice round mean sea level values, for example) although the file will need to have a column named MeanDepth.}
}
\description{
This function generates a series of map outputs based on the interval bins contained in the interval
file (see getintervals_time() and getintervals_sealvl()). This function takes a user-supplied input bathymetry raster and
generates a series of island extents based on the mean sea levels specified in the interval file. This function generates maps in
three formats: ESRI shapefile format, a flat raster format (with no topography), and a topographic raster format that preserves
the original bathymetric elevations of each island pixel, reprojected to the user-specified projected coordinate system.
}
\examples{
\dontrun{
#generate maps of the Fiji Archipelago with the EPSG:3141 projection, with default interval file
makemaps(inputraster="FJ.asc",epsg=3141,intervalfile="output/intervals.csv")
#if using a custom/modified interval file
makemaps(inputraster="FJ.asc",epsg=3141,intervalfile="output/customintervals/csv")
}
}
