% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spratt_lisecki_2016}
\alias{spratt_lisecki_2016}
\title{Alternative Sea Level Reconstruction Dataset (Spratt & Lisecki, 2016)}
\format{
\subsection{\code{spratt_lisecki_2016}}{

A data frame with 800 rows and 2 columns:
\describe{
\item{Time}{Time in the past (kya)}
\item{Sealevel_Corrected}{Sea level estimate, with present day set to 0
(metres)}
}
}
}
\source{
\url{https://doi.org/10.5194/cp-12-1079-2016}
}
\usage{
spratt_lisecki_2016
}
\description{
Sea level reconstruction file from Spratt & Lisecki (2016)
}
\keyword{datasets}
