% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewshed.R
\name{viewTo}
\alias{viewTo}
\title{Calculates if matrix of points is visible from another fixed point, given local topography
Code written by Barry Rowlingson (spacedman), posted on
https://stackoverflow.com/questions/21841387/r-code-that-evaluates-line-of-sight-los-between-two-lat-lon-points}
\usage{
viewTo(r, xy, xy2, h1 = 0, h2 = 0, progress = "none")
}
\description{
Calculates if matrix of points is visible from another fixed point, given local topography
Code written by Barry Rowlingson (spacedman), posted on
https://stackoverflow.com/questions/21841387/r-code-that-evaluates-line-of-sight-los-between-two-lat-lon-points
}
\keyword{internal}
