% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getintervals.R
\name{getintervals_time}
\alias{getintervals_time}
\title{Generate interval file, binning by time}
\usage{
getintervals_time(time, intervals, sealvl = bintanja_vandewal_2008)
}
\arguments{
\item{time}{Maximum time cutoff in kya. PleistoDist calculates the average distance between islands over a user-defined period of time. Users will have to specify an upper time bound (in thousands of years \link{kya}) for their PleistoDist analysis, which by default ranges from 0.1 kya to 3000 kya (i.e. 100 to 3,000,000 years ago). The lower time bound is fixed at the present day (0 kya).}

\item{intervals}{Number of time interval bins to generate between the present day and the time-cutoff. More time intervals will lead to more fine-grained geomorphology calculations, at the expense of increased computational time. Under default conditions, the number of intervals should not exceed time-cutoff*10.}

\item{sealvl}{Optional parameter defining the sea level reconstruction file. By default, PleistoDist uses the sea level reconstruction from Bintanja & van de Wal (2008). If you use a custom sea level reconstruction file, make sure the file has a Time column (in kya) and a Sealevel_Corrected column (in metres), and is saved in CSV format.}
}
\value{
An interval file in the automatically-generated 'output' folder. This interval file is needed for all downstream PleistoDist analyses
}
\description{
This function generates an interval file, which is the master file for controlling all downstream PleistoDist calculations. This file essentially simplifies Pleistocene-era sea level change into a number of discrete time intervals.
}
\examples{
#if using the default sea level reconstruction
getintervals_time(100,10) #time cutoff of 100,000 years, for 10 time intervals
#if using a custom sea level reconstruction
sealvlfile <- read.csv("reconstruction.csv")
getintervals_time(100,10,sealvl=sealvlfile) #time cutoff of 100,000 years, for 10 time intervals, using a custom sea level reconstruction file.
}
