% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcmetrics.R
\name{pleistodist_centroid}
\alias{pleistodist_centroid}
\title{Calculate inter-island centroid-to-centroid distance over time}
\usage{
pleistodist_centroid(points, epsg, intervalfile = "output/intervals.csv")
}
\arguments{
\item{points}{A user-generated multi-point shapefile (.shp) containing at least two points. If the shapefile attribute table contains a column
labelled 'Name', the output distance matrix will use the identifiers in this column to label each pairwise comparison. Otherwise, the
output distance matrix will use the attribute FID values instead.}

\item{epsg}{The projected coordinate system in EPSG code format. Because of the curvature of the Earth's surface, we need to apply a map
projection to accurately calculate straight-line distances between points instead of using the default WGS84 geographical coordinate system.
Users should specify a projected coordinate system appropriate to the geographic region being analysed using the projection's
associated EPSG code (https://epsg.org/home). Geographic coordinate system projections are not recommended as those will result
in distance matrices calculated in decimal degrees rather than in distance units.}

\item{intervalfile}{This is the master control file generated using either the getintervals_time() or getintervals_sealvl()
function that defines the number of intervals, the sea level at each interval, and the duration of each interval. By default, this
function will use the "intervals.csv" file stored in the output folder, but users can also specify their own custom interval
file (with nice round mean sea level values, for example), although users need to ensure that the same column names are preserved, and
be aware that custom interval files may lead to inaccurate weighted mean distance calculations.}
}
\value{
This function outputs a pairwise distance matrix of pairwise inter-island centroid-to-centroid distances in long format, with
one column per interval.
}
\description{
This function calculates the centroid-to-centroid distance between island landmasses over Pleistocene time. For each pairwise combination
of source points provided by the user, this function selects the landmasses that correspond to each point, and calculates the distance between
the centroids of the selected landmasses for each time/sea level interval. If both points lie on the same landmass, a distance value of 0 is returned. If one or both islands
is/are underwater, then a value of NA is returned. After calculating the pairwise island distances for each interval, this function calculates
the time-weighted average centroid-to-centroid distance between each island pair.
}
