% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcmetrics.R
\name{pleistodist_netmig}
\alias{pleistodist_netmig}
\title{Estimate net inter-island migration over time}
\usage{
pleistodist_netmig(
  points,
  epsg,
  disttype,
  intervalfile = "output/intervals.csv"
)
}
\arguments{
\item{points}{A user-generated multi-point shapefile (.shp) containing at least two points. If the shapefile attribute table contains a column
labelled 'Name', the output distance matrix will use the identifiers in this column to label each pairwise comparison. Otherwise, the
output distance matrix will use the attribute FID values instead.}

\item{epsg}{The projected coordinate system in EPSG code format. Because of the curvature of the Earth's surface, we need to apply a map
projection to accurately calculate straight-line distances between points instead of using the default WGS84 geographical coordinate system.
Users should specify a projected coordinate system appropriate to the geographic region being analysed using the projection's
associated EPSG code (https://epsg.org/home). Geographic coordinate system projections are not recommended as those will result
in distance matrices calculated in decimal degrees rather than in distance units.}

\item{disttype}{The type of inter-island distance matrix to use for this calculation, either "centroid", "leastshore", or "meanshore".}

\item{intervalfile}{This is the master control file generated using either the getintervals_time() or getintervals_sealvl()
function that defines the number of intervals, the sea level at each interval, and the duration of each interval. By default, this
function will use the "intervals.csv" file stored in the output folder, but users can also specify their own custom interval
file (with nice round mean sea level values, for example), although users need to ensure that the same column names are preserved, and
be aware that custom interval files may lead to inaccurate weighted mean calculations.}
}
\value{
This function outputs a matrix of island surface area estimates in long format, with one column per interval.
}
\description{
This function makes use of the equation from chapter 6 of MacArthur & Wilson (1967) that estimates the number of propagules from a
source island arriving at a receiving island as a function of inter-island distance, area, and relative width. Although estimating
the unidirectional rate of migration from one island to another requires several other coefficients (alpha: the coefficient of population density,
and lambda: the coefficient of overwater dispersal), calculating the ratio of migration rate for bidirectional migration causes
these coefficients to cancel out, allowing for net migration rate to be estimated using purely geomorphological and distance-based
metrics. The net migration rate can therefore be calculated using the equation (atan(t2/(2*d))\emph{A1)/(atan(t1/(2}d))*A2), where t is
the island width relative to the other island, d is the inter-island distance, and A is the island area, for islands 1 and 2. If the
net migration rate is greater than 1, then the net movement of propagules occurs from island 1 to island 2. If the net migration rate
is less than 1, then the net movement of propagules occurs from island 2 to island 1.
}
